clean bookshelf code, add DropDownItems type
