+++ src/components/types.ts
@@ -1,0 +1,4 @@
+export type DropDownItems = {
+  placeholder: string;
+  menuItems: {label: string; value: string}[];
+};
