+++ src/screens/Bookshelf.tsx
@@ -1,3 +1,3 @@
-import { BookshelfSectionList } from '@src/feature/BookshelfSectionList';
+import { BookshelfSectionList } from '@src/feature/BookshelfSectionList/BookshelfSectionList';
 
 export const Bookshelf = () => <BookshelfSectionList />;
