+++ src/screens/Home.tsx
@@ -1,7 +1,7 @@
 import {View} from 'react-native';
 import {bookmarkData, wishlistData} from '@src/data/MockData';
 import {styles} from './styles';
-import {HomeScreenFlatList} from '@src/feature/HomeScreenFlatList';
+import {HomeScreenFlatList} from '@src/feature/HomeFlatListSection/HomeScreenFlatList';
 import {Text} from '@components/Text';
 
 // The same data should be able to be passed in without edit, adjustments to data should be handled within component
